Move the request routing logic to the empty servlet.  Remove old PowerFilterChain class and fix minor typo in web.xml comment.
