+++ project-set/core/core-lib/src/main/java/com/rackspace/papi/filter/RequestFilterChainState.java
@@ -1,6 +1,5 @@
 package com.rackspace.papi.filter;
 
-import com.rackspace.papi.commons.util.StringUtilities;
 import javax.servlet.FilterChain;
 import javax.servlet.ServletException;
 import javax.servlet.ServletRequest;
@@ -8,9 +7,6 @@ import javax.servlet.ServletResponse;
 import java.io.IOException;
 import java.util.LinkedList;
 import java.util.List;
-import javax.servlet.RequestDispatcher;
-import javax.servlet.ServletContext;
-import javax.servlet.http.HttpServletRequest;
 
 /**
  * @author fran
@@ -30,19 +26,16 @@ public class RequestFilterChainState implements javax.servlet.FilterChain {
     private final List<FilterContext> filterChainCopy;
     private final FilterChain containerFilterChain;
     private final ClassLoader containerClassLoader;
-    private final ServletContext context;
     private int position;
 
-    public RequestFilterChainState(List<FilterContext> filterChainCopy, FilterChain containerFilterChain, ServletContext context) {
+    public RequestFilterChainState(List<FilterContext> filterChainCopy, FilterChain containerFilterChain) {
         this.filterChainCopy = new LinkedList<FilterContext>(filterChainCopy);
         this.containerFilterChain = containerFilterChain;
-        this.context = context;
         this.containerClassLoader = Thread.currentThread().getContextClassLoader();
     }
     
     public void startFilterChain(ServletRequest servletRequest, ServletResponse servletResponse) throws IOException, ServletException {
         doFilter(servletRequest, servletResponse);
-        route(servletRequest, servletResponse);
     }
 
     @Override
@@ -71,19 +64,6 @@ public class RequestFilterChainState implements javax.servlet.FilterChain {
                 containerFilterChain.doFilter(servletRequest, servletResponse);
             } finally {
                 currentThread.setContextClassLoader(previousClassLoader);
-            }
-        }
-    }
-
-    private void route(ServletRequest servletRequest, ServletResponse servletResponse) throws IOException, ServletException {
-        final String uriPath = ((HttpServletRequest) servletRequest).getRequestURI();
-
-        if (!StringUtilities.isBlank(uriPath)) {
-            final ServletContext targetContext = context.getContext(uriPath);
-            
-            if (targetContext != null) {
-                final RequestDispatcher dispatcher = targetContext.getRequestDispatcher(uriPath);
-                dispatcher.forward(servletRequest, servletResponse);
             }
         }
     }
